% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_stats}
\alias{cache_stats}
\title{Report statistics on cache utilization.}
\usage{
cache_stats(fn)
}
\arguments{
\item{fn}{A memoized function (such as created by \code{\link{memo}})}
}
\value{
A list with entries "hits", "misses", "expired" and "entries"
counting how many times the cache has found a previously saved result,
computed a new result, and dropped an old result from memory, and the number
of entries currently in the cache.
}
\description{
Report statistics on cache utilization.
}

