% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cache.R
\name{digest_key}
\alias{digest_key}
\title{Memoize based on object serialization.}
\usage{
digest_key(x)
}
\arguments{
\item{x}{A list of objects.}
}
\value{
A string based on a hash of the object contents.
}
\description{
This is the default key type for \code{\link{memo}}.
}

