% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cache.R, R/memo-description.r
\docType{package}
\name{memo}
\alias{memo}
\alias{memo-package}
\title{Memoize a function based on the raw pointers of its arguments.}
\usage{
memo(fn, cache = lru_cache(1000), key = digest_key)
}
\arguments{
\item{fn}{A function to wrap. It should be a pure function (i.e. it
should not cause side effects, and should not depend on any
variables that may change.)}

\item{cache}{A cache to use. Defaults to a new instance of \code{\link{lru_cache}}.
Caches may be shared between memoized functions.}
}
\description{
Memoize a function based on the raw pointers of its arguments.

This package implements a cache that can be used to avoid repeated
computations of functions. The cache lookup is based on object
identity (i.e.  pointer equivalence) which is suited for functions
like accessors or other functions that are called repeatedly on the
same object.  Description of memo goes here.
}
\author{
Peter Meilstrup
}

