% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cache.R
\name{pointer_key}
\alias{pointer_key}
\title{Memoize based on pointer equivalence.}
\usage{
pointer_key(x)
}
\arguments{
\item{x}{A list of objects.}
}
\value{
A string encoding the memory locations of each item in the
list.
}
\description{
This may be used as an argument to 'memo' to use pointer
equivalence instead of a computed hash to index the cache.
}

