% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_stats}
\alias{cache_stats}
\title{Report cache statistics.}
\usage{
cache_stats(fn)
}
\arguments{
\item{fn}{A memoized function that was created by \code{\link{memo}}.}
}
\value{
A list with labels "size", "used", "hits", "misses", "expired"
counting the number of slots in the cache, the number of slots currently
used, the number of times a previous result was recalled, a new result was
recorded, and a result was dropped.
}
\description{
Report cache statistics.
}
