% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lru.R
\name{lru_cache}
\alias{lru_cache}
\title{A cache keyed on strings with LRU policy.}
\usage{
lru_cache(size = 1000)
}
\arguments{
\item{size}{The maximum number of results to keep.}
}
\value{
A function f(key, value) which takes a string in the first
parameter and a lazily evaluated value in the second. The function
will return the matching from the cache, or force the second
argument and return that, remembering the result on future calls.

When the number of entries in the cache exceeds \code{size}, the least
recently accessed entries are removed.
}
\description{
A cache keyed on strings with LRU policy.
}

