% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lru.R
\name{lru_cache}
\alias{lru_cache}
\title{Construct a cache with least-recently-used policy.}
\usage{
lru_cache(size = 10000)
}
\arguments{
\item{size}{The maximum number of results to keep.}
}
\value{
A function f(key, value) which takes a string in the first
parameter and a lazily evaluated value in the second. `f`
will use the string key to retrieve a value from the cache, or 
return the matching item from the cache, or force the second
argument and return that, remembering the result on future calls.

When the number of entries in the cache exceeds \code{size}, the least
recently accessed entries are removed.
}
\description{
Construct a cache with least-recently-used policy.
}
