% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memo-description.r
\docType{package}
\name{memo-package}
\alias{memo-package}
\alias{_PACKAGE}
\title{In-memory caching of repeated computations, by pointer equivalence.}
\description{
The `memo` package implements a cache that can be used to avoid repeated
computations of functions. The cache lookup is based on object
identity (i.e.  pointer equivalence) which is suited for functions
like accessors or other functions that are called repeatedly on the
same object.
}
\author{
Peter Meilstrup
}
