% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cache.R, R/memo-description.r
\docType{package}
\name{memo}
\alias{memo}
\alias{memo-package}
\title{Memoize a function based on the raw pointers of its arguments.}
\usage{
memo(fn, cache = global_cache)
}
\description{
Memoize a function based on the raw pointers of its arguments.

Description of memo goes here.
}
\author{
Peter Meilstrup
}

