% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{strategies}
\alias{strategies}
\alias{digest_key}
\alias{pointer_key}
\alias{hybrid_key}
\title{Strategies for caching items.}
\usage{
digest_key(fn, cache, digest = digest::digest)

pointer_key(fn, cache)

hybrid_key(fn, cache, digest = function(x) digest::digest(x, "md5"))
}
\arguments{
\item{fn}{A function whose results should be cached.}

\item{cache}{A cache object.}

\item{digest}{A digest function to use.}
}
\value{
A memoized function.
}
\description{
The function \code{\link{memo}} accepts an argument `key` which
specifies the keying strategy.
}
